Move handlers to their own files and read config from default location
